requirements.txt
python-chess
numpy
cython
